refactor: use context path to reduce DB queries
and do not catch errors
